refactor(go): extract l2topology OUI lookup
